live-flask-starter/requirements.txt
deepgram-sdk==v3.4.0
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
qdrant-client==1.10.1